backend/requirements.txt
django~=4.1.5
djangorestframework
psycopg2
djangorestframework_simplejwt
shortuuid
channels
daphne